// front/src/common/enums/selectWithButton.ts
enum TEXT_ENUM {
  CHOOSE_SORT = 'Выберите вид сортировки',
  ADD_PRODUCT = 'Добавить чебурек',
}

enum SORT_ENUM {
  DATE_CREATED_AT = "По дате создания в базе данных",
  DATE_END = 'По дате окончания срока годностии',
  DATE_START = 'По дате изготовления',
}

export { TEXT_ENUM, SORT_ENUM };